backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
pymongo==4.6.0
pydantic==2.5.0
python-dateutil==2.8.2
httpx==0.25.2
python-dotenv==1.0.0